api/requirements.txt
Django==4.2
psycopg2-binary==2.8.6
djangorestframework==3.14.0
celery==5.3.1
requests==2.31.0
django-cors-headers==4.2.0